+++ awslimitchecker/runner.py
@@ -133,6 +133,11 @@ class Runner(object):
                        type=int, default=99,
                        help='default critical threshold (percentage of '
                        'limit); default: 99')
+        p.add_argument('-P', '--profile', action='store', dest='profile_name',
+                       type=str, default=None,
+                       help='Name of profile in the AWS cross-sdk credentials '
+                            'file to use credentials from; similar to the '
+                            'corresponding awscli option')
         p.add_argument('-A', '--sts-account-id', action='store',
                        type=str, default=None,
                        help='for use with STS, the Account ID of the '
@@ -311,6 +316,7 @@ class Runner(object):
         self.checker = AwsLimitChecker(
             warning_threshold=args.warning_threshold,
             critical_threshold=args.critical_threshold,
+            profile_name=args.profile_name,
             account_id=args.sts_account_id,
             account_role=args.sts_account_role,
             region=args.region,